feat: Services classes added and Projeto model added, the templates were changed to receive this changes and validation made but it isn't ready

+++ src/main/java/voyager/people_management/repositories/CargoRepository.java
@@ -3,5 +3,5 @@ package voyager.people_management.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import voyager.people_management.models.Cargo;
 
-public interface CargoRepository extends JpaRepository<Cargo, Integer> {
+public interface CargoRepository extends JpaRepository<Cargo, Long> {
 }
