+++ src/main/java/voyager/people_management/repositories/DepartamentoRepository.java
@@ -3,5 +3,5 @@ package voyager.people_management.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import voyager.people_management.models.Departamento;
 
-public interface DepartamentoRepository extends JpaRepository<Departamento, Integer> {
+public interface DepartamentoRepository extends JpaRepository<Departamento, Long> {
 }
